Join an email list
Prompt: Create a form that has four fields. If the email address entered first does not match the reentry, show an error and do not let user continue. If any of the fields are left empty, prompt an error. If no errors are found, accept submission and return a thank you email (attached in separate html file.