Create a Bill Calculator using Javascript
Code (without attachments) is listed first with HTML5, followed by the Javascript and ends with CSS formatting.

Prompt: Create a Bill Calculator. Provide errors if the first or second fields are empty. Also, if a user enters a value greater than 1000 in the "Subtotal" field, return an error prompting the user to enter a number 0-1000. Additionally, show an error if the user does not enter a number 0-25 in the "Tax Rate" field. The "Sales Tax" and "Total" fields must not be manipulated by the user. If no errors are presented and the form passes all rules allow the user to calculate the Sales Tax and return the "Total" to the user. Allow for a clear function that the user can clear all fields.